Removed sample legal text
