+++ src/app/components/App/App.js
@@ -11,9 +11,8 @@ import { HomePage } from '../../../home';
 import { CategoriesPage } from '../../../categories';
 import { ProposalsPage } from '../../../proposals';
 import { InstitutionsPage } from '../../../institutions';
-import Terms from '../footer/Terms';
-import PrivacyPolicy from '../footer/PrivacyPolicy';
-import Information from '../footer/Information';
+import Terms from '../footer/legal/Terms';
+import PrivacyPolicy from '../footer/legal/PrivacyPolicy';
 
 import Footer from '../footer/Footer';
 
@@ -57,7 +56,6 @@ class App extends Component {
                 <Route path="/institutions" component={InstitutionsPage} />
                 <Route path="/terms" component={Terms} />
                 <Route path="/privacy" component={PrivacyPolicy} />
-                <Route path="/information" component={Information} />
               </div>
             </div>
             <Footer />
